requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
Baker==1.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
django-redis==5.4.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==23.1.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
model-bakery==1.17.0
packaging==23.2
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
